modularizing question cell; created general question delegate
